% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steam.R
\name{steam_app}
\alias{steam_app}
\title{Get detailed information for a Steam app}
\usage{
steam_app(appid)
}
\arguments{
\item{appid}{Steam application ID (e.g. 730 for CS:GO, 570 for Dota 2)}
}
\value{
tibble with one row: appid, name, developer, publisher, positive,
negative, owners, average_forever, average_2weeks, median_forever,
median_2weeks, price, initialprice, discount, ccu, genre, languages
}
\description{
Get detailed information for a Steam app
}
