% Generated by build script
\name{steam_genre}
\alias{steam_genre}
\title{Get games by genre}
\usage{
steam_genre()
}
\arguments{
\item{genre}{Genre name (e.g. "Action", "RPG", "Strategy", "Indie",}
}
\value{tibble: appid, name, developer, publisher, positive, negative,}
\description{Get games by genre}
