% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steam.R
\name{steam_top100}
\alias{steam_top100}
\title{Get top 100 games by current players in last 2 weeks}
\usage{
steam_top100()
}
\value{
tibble: appid, name, developer, publisher, positive, negative,
owners, average_forever, average_2weeks, price, initialprice, discount, ccu
}
\description{
Get top 100 games by current players in last 2 weeks
}
