% Generated by build script
\name{steam_app}
\alias{steam_app}
\title{Get detailed information for a Steam app}
\usage{
steam_app()
}
\arguments{
\item{appid}{Steam application ID (e.g. 730 for CS:GO, 570 for Dota 2)}
}
\value{tibble with one row: appid, name, developer, publisher, positive,}
\description{Get detailed information for a Steam app}
