% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steam.R
\name{steam_genre}
\alias{steam_genre}
\title{Get games by genre}
\usage{
steam_genre(genre)
}
\arguments{
\item{genre}{Genre name (e.g. "Action", "RPG", "Strategy", "Indie",
"Adventure", "Simulation", "Racing", "Sports")}
}
\value{
tibble: appid, name, developer, publisher, positive, negative,
owners, average_forever, average_2weeks, price, initialprice, discount, ccu
}
\description{
Get games by genre
}
